+++ classes/Email.php
@@ -136,12 +136,16 @@ class Email
     {
         /** @var Twig $twig */
         $twig = Grav::instance()['twig'];
+        $twig->init();
 
         /** @var Config $config */
         $config = Grav::instance()['config'];
 
         /** @var Language $language */
         $language = Grav::instance()['language'];
+
+        // Create message object.
+        $message = $this->message();
 
         // Extend parameters with defaults.
         $params += [
@@ -159,11 +163,9 @@ class Email
             'to' => $config->get('plugins.email.to'),
             'to_name' => $config->get('plugins.email.to_name'),
             'process_markdown' => false,
-            'template' => false
+            'template' => false,
+            'message' => $message
         ];
-
-        // Create message object.
-        $message = $this->message();
 
         if (!$params['to']) {
             throw new \RuntimeException($language->translate('PLUGIN_EMAIL.PLEASE_CONFIGURE_A_TO_ADDRESS'));
@@ -233,7 +235,9 @@ class Email
                     break;
 
                 case 'subject':
-                    $message->setSubject($twig->processString($language->translate($value), $vars));
+                    if ($value) {
+                        $message->setSubject($twig->processString($language->translate($value), $vars));
+                    }
                     break;
 
                 case 'to':